+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useDebouncedValue';
+export * from './useFieldColor';
 export * from './useFormValues';
 export * from './useOutsideClick';
 export * from './useTypeaheadQuery';